HIVE-27152: Revert "HIVE-21388: Constant UDF is not pushed to JDBCStorage Handler" (Aman Raj, reviewed by Vihang Karajgaonkar)
This reverts commit 008797cb2fab6291197459fddce35d61e638c004.
