+++ ql/src/java/org/apache/hadoop/hive/ql/optimizer/calcite/HiveCalciteUtil.java
@@ -69,7 +69,6 @@ import org.apache.hadoop.hive.ql.exec.FunctionRegistry;
 import org.apache.hadoop.hive.ql.metadata.VirtualColumn;
 import org.apache.hadoop.hive.ql.optimizer.calcite.reloperators.HiveMultiJoin;
 import org.apache.hadoop.hive.ql.optimizer.calcite.reloperators.HiveProject;
-import org.apache.hadoop.hive.ql.optimizer.calcite.reloperators.HiveSqlFunction;
 import org.apache.hadoop.hive.ql.optimizer.calcite.reloperators.HiveTableFunctionScan;
 import org.apache.hadoop.hive.ql.optimizer.calcite.translator.ExprNodeConverter;
 import org.apache.hadoop.hive.ql.optimizer.calcite.translator.SqlFunctionConverter;
@@ -979,6 +978,7 @@ public class HiveCalciteUtil {
    * Check if the expression is usable for query materialization, returning the first failing expression.
    */
   public static RexCall checkMaterializable(RexNode expr) {
+    boolean deterministic = true;
     RexCall failingCall = null;
 
     if (expr == null) {
@@ -989,9 +989,7 @@ public class HiveCalciteUtil {
       @Override
       public Void visitCall(org.apache.calcite.rex.RexCall call) {
         // non-deterministic functions as well as runtime constants are not materializable.
-        SqlOperator op = call.getOperator();
-        if (!op.isDeterministic() || op.isDynamicFunction() ||
-            (op instanceof HiveSqlFunction && ((HiveSqlFunction) op).isRuntimeConstant())) {
+        if (!call.getOperator().isDeterministic() || call.getOperator().isDynamicFunction()) {
           throw new Util.FoundOne(call);
         }
         return super.visitCall(call);
