+++ ql/src/test/queries/clientpositive/current_date_timestamp.q
@@ -2,8 +2,6 @@ select current_timestamp = current_timestamp(), current_date = current_date() fr
 
 set hive.test.currenttimestamp =2012-01-01 01:02:03;
 
-explain cbo select current_timestamp() from alltypesorc;
-
 --ensure that timestamp is same for all the rows while using current_timestamp() query should return single row
 select count(*) from (select current_timestamp() from alltypesorc union select current_timestamp() from src limit 5 ) subq;
